Phase 2c-2A: pipeline buffers unsigned char -> uint16 (scale still 0-255)
bufferFaders, bufferSequenciel (channels) and MergerArray (patch) are now
unsigned short[514], holding plain 0-255 values for now -> strictly identical
behaviour (non-regression validated). Only two Tmax calls needed a cast
(template is single-type; ushort buffer vs uchar FaderDoDmx); no block/pointer
usage to adapt. Sets the stage for the x257 16-bit scale (2c-2B).

+++ Src/dmx_functions.cpp
@@ -819,12 +819,12 @@ int Merger_Faders() {
             case 0: // HTP normal
                 if (!fader_fx_route[cif])
                 {
-                    bufferFaders[h] = Tmax(bufferFaders[h], FaderDoDmx[cif][h]);
+                    bufferFaders[h] = Tmax(bufferFaders[h], (unsigned short)FaderDoDmx[cif][h]);
                 }
                 else
                 {
                     if (FaderDoDmx[cif][h] > 0) {
-                        bufferSequenciel[h] = Tmax(bufferSequenciel[h], FaderDoDmx[cif][h]);
+                        bufferSequenciel[h] = Tmax(bufferSequenciel[h], (unsigned short)FaderDoDmx[cif][h]);
                         channel_is_touched_by_fader_fx[h] = 1;
                         channel_is_touched_by_fader_number[h] = cif;
                         channel_is_touched_by_fader_type_fx[h] = fader_mode_with_buffers[cif];
